+++ R/sim_mixed_cc.R
@@ -8,6 +8,7 @@
 #' @param sub_sd the SD of subject random intercepts
 #' @param item_sd the SD of item random intercepts
 #' @param error_sd the SD of the error term
+#' @param empirical Should the returned data have these exact parameters? (versus be sampled from a population with these parameters)
 #' @param seed a single value, interpreted as an integer, or NULL (see set.seed)
 #'
 #' @return a tbl 
@@ -17,7 +18,8 @@
 #' 
 #' sim_mixed_cc(10, 10)
 sim_mixed_cc <- function(sub_n = 100, item_n = 20, grand_i = 0, 
-                         sub_sd = 1, item_sd = 1, error_sd = 1, seed = NULL) {
+                         sub_sd = 1, item_sd = 1, error_sd = 1, 
+                         empirical = FALSE, seed = NULL) {
   if (!is.null(seed)) {
     # reinstate system seed after simulation
     sysSeed <- .GlobalEnv$.Random.seed
@@ -42,7 +44,9 @@ sim_mixed_cc <- function(sub_n = 100, item_n = 20, grand_i = 0,
     # sample new subjects and intercepts
     new_sub <- data.frame(
       sub_id = make_id(1:sub_n, "S"),
-      sub_i = stats::rnorm(sub_n, 0, sub_sd)
+      sub_i = rnorm_multi(sub_n, 1, 0, sub_sd, 
+                          empirical = empirical, 
+                          as.matrix = TRUE)[,1]
     )
   }
   
@@ -58,7 +62,9 @@ sim_mixed_cc <- function(sub_n = 100, item_n = 20, grand_i = 0,
     # sample new items and intercepts
     new_item <- data.frame(
       item_id = make_id(1:item_n, "I"),
-      item_i = stats::rnorm(item_n, 0, item_sd)
+      item_i = rnorm_multi(item_n, 1, 0, item_sd, 
+                           empirical = empirical, 
+                           as.matrix = TRUE)[,1]
     )
   }